fix(computers): honest cloud-only unconfigured message, shared component
The dataPlaneUnavailable empty state told every user — including OSS users
who can't act on it — to set COMPUTERS_REMOTE_DATA_PLANE_URL or the
data-plane secrets. Computers are cloud-only: an inspector either is a
deployed data plane or delegates to the one its Convex deployment
advertises, so the honest message is that the connected MCPJam deployment
hasn't enabled Project Computers.

Extracts the twice-duplicated message into ComputersUnavailableMessage and
the twice-duplicated PaneMessage into its own module (ComputerView carried
a byte-identical local copy).
